requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
openai
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
